Loading Refactor
